camera movement is set up

+++ include/InputManager.hpp
@@ -14,7 +14,7 @@ class InputManager
 		~InputManager(void);
 		void	addObserver(InputObserver*);
 		void	eraseObserver(InputObserver*);
-		void	notifyOnKeyDown(SDL_Scancode, double);
+		void	notifyOnKeyDown(SDL_Scancode, float);
 		void	notifyOnMouseMove(Uint8);
 	private:
 		std::vector<InputObserver*> observers;
